Requirement (feat) pause song - Extracted utility functions into a separate file. - Moved component logic into custom hooks for better separation of concerns. - Split the component into smaller components for easier management and maintainability. - a new restartSong function to the MediaPlayer component that allows the user to restart the current song from the beginning.
