+++ src/utils/colors.ts
@@ -1,0 +1,10 @@
+const colors = {
+  ORANGE_COLOR: '#ff6600',
+  GRAY_COLOR: '#cccccc',
+  LIGHT_GRAY_COLOR: '#e5e5e5',
+  GREEN_COLOR: '#1DB954',
+  GRAY_COLOR_VARIATION: '#e1e1e1',
+  DARK_GRAY_COLOR: '#999',
+};
+
+export default colors;
